+++ server/functions/pingHelper.js
